doctalk/requirements.txt
beautifulsoup4==4.12.2
datasets==2.16.1
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-pinecone==0.0.3
langchain-text-splitters==0.0.2
octoai-sdk==0.10.1
openai==1.30.1
pinecone-datasets==0.7.0
pinecone-client==3.2.2
playwright==1.40.0
python-dotenv==1.0.0
tiktoken==0.5.2
transformers==4.37.2
torch==2.2.0